Don't hide controllers behind macros
It was fun to have controllers completely hidden, so that everything was a 0-32768 range of values that you had no clue as to how it connected to the synth.

However, they were not 100% bug free in operation, and would sometimes cause nice-sounding modules to lose their original form after a certain modulation, never to return again.